moved helpers, removed css
