+++ app/assets/stylesheets/application.sass
@@ -141,15 +141,6 @@ table.custom_striped.details
 td.truncate
   max-width: 0
 
-.protocols-group
-  .truncate
-    width: 29em !important
-
-.truncate
-  white-space: nowrap
-  overflow: hidden
-  text-overflow: ellipsis
-
 div.tooltip-inner
   max-width: 800px
